+++ R/g32-data.R
@@ -7,7 +7,7 @@
 #' \describe{
 #' \item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
 #' \item{\code{occupy}}{An indicator for whether a dwelling is occupied or not}
-#' \item{\code{dwelling type}}{A description of the type of dwelling}
+#' \item{\code{dwelling_type}}{A description of the type of dwelling}
 #' \item{\code{storeys}}{The number of storeys for semi detached structures,
 #' flats and apartments}
 #' \item{\code{dwellings}}{The count of dwellings}